Added Analysis Command
