+++ deepview_profile/__main__.py
@@ -8,7 +8,7 @@ import deepview_profile
 import deepview_profile.commands.interactive
 import deepview_profile.commands.memory
 import deepview_profile.commands.time
-import deepview_profile.commands.complete
+import deepview_profile.commands.analysis
 
 
 def main():
@@ -26,7 +26,7 @@ def main():
     deepview_profile.commands.interactive.register_command(subparsers)
     deepview_profile.commands.memory.register_command(subparsers)
     deepview_profile.commands.time.register_command(subparsers)
-    deepview_profile.commands.complete.register_command(subparsers)
+    deepview_profile.commands.analysis.register_command(subparsers)
     args = parser.parse_args()
 
     if args.version:
